fix изменения коннекторов

+++ Traffic_project/src/main/java/ru/nsu/fit/traffic/network/ConnectionImpl.java
@@ -67,7 +67,7 @@ public class ConnectionImpl implements Connection {
   }
 
   @Override
-  public void pushGlobalMap(String filePath, Long roomId) {
+  public void pushGlobalMap(String filePath, Integer roomId) {
     pushAnyMap(filePath, SAVE_URL + "?roomId=" + roomId);
   }
