+++ Traffic_project/src/main/java/ru/nsu/fit/traffic/network/stub/ConnectionStub.java
@@ -45,7 +45,7 @@ public class ConnectionStub implements Connection {
   }
 
   @Override
-  public void pushGlobalMap(String filePath, Long roomId) {
+  public void pushGlobalMap(String filePath, Integer roomId) {
 
   }
